// backend/src/common/enums/role.enum.ts
export enum Role {
    SUPER_ADMIN = 'SUPER_ADMIN',
    UNIVERSITY_ADMIN = 'UNIVERSITY_ADMIN',
    REGISTRAR = 'REGISTRAR',
    HOD = 'HOD',
    ACADEMIC_COORDINATOR = 'ACADEMIC_COORDINATOR',
    FACULTY = 'FACULTY',
    STUDENT = 'STUDENT',
    EXAM_CONTROLLER = 'EXAM_CONTROLLER',
    FINANCE = 'FINANCE',
    ACCOUNTANT = 'ACCOUNTANT',
    LIBRARIAN = 'LIBRARIAN',
    HOSTEL_WARDEN = 'HOSTEL_WARDEN',
    PLACEMENT_CELL = 'PLACEMENT_CELL',
    PLACEMENT_OFFICER = 'PLACEMENT_OFFICER',
}